Optimized linting
