+++ .eslintrc
@@ -11,7 +11,13 @@
         "browser": true,
         "node": true
     },
-    "extends": "eslint:recommended",
+    "extends": [
+        "eslint:recommended",
+        "plugin:import/errors",
+        "plugin:import/warnings",
+        "plugin:lodash/recommended",
+        "plugin:react/recommended"        
+    ],
     "parser": "babel-eslint",
     "rules": {
         "quotes": [
@@ -22,17 +28,26 @@
             2,
             "never"
         ],
+        "no-var": 2,
+        "prefer-spread": 2,
+        "prefer-const": 2,
+        "prefer-template": 2,
         "babel/generator-star-spacing": 1,
         "babel/new-cap": 1,
         "babel/object-shorthand": 1,
         "babel/arrow-parens": 1,
         "babel/no-await-in-loop": 1,
-        "react/jsx-uses-react": 2,
-        "react/jsx-uses-vars": 2,
-        "react/react-in-jsx-scope": 2
+        "react/no-multi-comp": [
+            2,
+            {
+                "ignoreStateless": true
+            }
+        ]
     },
     "plugins": [
         "babel",
+        "import",
+        "lodash",
         "react"
     ]
 }
