+++ package.json
@@ -43,7 +43,9 @@
     "cross-env": "^1.0.7",
     "eslint": "^2.4.0",
     "eslint-plugin-babel": "^3.1.0",
-    "eslint-plugin-react": "^4.2.3",
+    "eslint-plugin-import": "^1.4.0",
+    "eslint-plugin-lodash": "^1.6.3",
+    "eslint-plugin-react": "^4.3.0",
     "eventsource-polyfill": "^0.9.6",
     "express": "^4.13.4",
     "ghooks": "^1.2.0",
