+++ src/App.js
@@ -1,33 +1,8 @@
-import React, { Component } from 'react';
-import { NICE, SUPER_NICE } from './colors';
+import React, { Component } from 'react'
+import { NICE, SUPER_NICE } from './colors'
+import Counter from './Counter'
 
-class Counter extends Component {
-    constructor(props) {
-        super(props);
-        this.state = { counter: 0 };
-        this.interval = setInterval(() => this.tick(), 1000);
-    }
-
-    tick() {
-        this.setState({
-            counter: this.state.counter + this.props.increment
-        });
-    }
-
-    componentWillUnmount() {
-        clearInterval(this.interval);
-    }
-
-    render() {
-        return (
-            <h1 style={{ color: this.props.color }}>
-                Counter ({this.props.increment}): {this.state.counter}
-            </h1>
-        );
-    }
-}
-
-export class App extends Component {
+export default class App extends Component {
     render() {
         return (
             <div>
